Move backend libraries to common folder
