+++ .justfile
@@ -4,9 +4,9 @@ set dotenv-required := true
 workspace_folder := "."
 tasks_folder := workspace_folder + "/tasks"
 tasks_venv_folder := tasks_folder + "/.venv"
-backend_dir := workspace_folder + "/backend"
+backend_api_dir := workspace_folder + "/backend/backend_api"
 frontend_dir := workspace_folder + "/frontend"
-frb_dart_output_dir := frontend_dir + "/lib/backend"
+frb_dart_output_dir := frontend_dir + "/lib/backend_api"
 llvmPath := `clang -v 2>&1 | grep 'Selected GCC installation' | rev | cut -d' ' -f1 | rev`
 llvmIncludeDir := llvmPath + "/include"
 venv_python := tasks_venv_folder + "/bin/python"
@@ -38,24 +38,24 @@ database-drop-tables: _tasks-create-venv
 database-recreate-tables: database-drop-tables database-create-tables
 
 diesel-setup:
-    cd {{ backend_dir }} && diesel setup
+    cd {{ backend_api_dir }} && diesel setup
 
 diesel-generate-schema: diesel-setup
-    cd {{ backend_dir }} && diesel print-schema > src/schema.rs
+    cd {{ backend_api_dir }} && diesel print-schema > src/schema.rs
 
 diesel-generate-models: diesel-generate-schema
-    cd {{ backend_dir }} && diesel_ext --model --import-types diesel::Queryable --import-types diesel::Selectable --import-types diesel::Identifiable --import-types backend_macros::make_fields_non_final --import-types flutter_rust_bridge::frb --import-types crate::schema::* --derive Queryable,Selectable --add-table-name > src/models.rs
+    cd {{ backend_api_dir }} && diesel_ext --model --import-types diesel::Queryable --import-types diesel::Selectable --import-types diesel::Identifiable --import-types backend_macros::make_fields_non_final --import-types flutter_rust_bridge::frb --import-types crate::schema::* --derive Queryable,Selectable --add-table-name > src/models.rs
     git apply {{ patch_folder }}/backend/models.rs.patch
 
 # FIXME Verify that FRB versions in Cargo.toml, pubspec.yaml and the installed FRB codegen (locally and in Github
 # Actions) correspond to each other
 frb-generate: diesel-generate-models
     mkdir -p {{ frb_dart_output_dir }}
-    flutter_rust_bridge_codegen generate --no-web --no-add-mod-to-lib --llvm-path {{ llvmIncludeDir }} --rust-input "crate::api" --rust-root {{ backend_dir }} --dart-output {{ frb_dart_output_dir }}
+    flutter_rust_bridge_codegen generate --no-web --no-add-mod-to-lib --llvm-path {{ llvmIncludeDir }} --rust-input "crate::api" --rust-root {{ backend_api_dir }} --dart-output {{ frb_dart_output_dir }}
     git apply {{ patch_folder }}/frontend/models.dart.patch
 
 backend-build: frb-generate
-    cd {{ backend_dir }} && cargo build --release
+    cd {{ backend_api_dir }} && cargo build --release
 
 frontend-generate-reflectable: frb-generate
     cd {{ frontend_dir }} && dart run build_runner build --delete-conflicting-outputs
@@ -71,7 +71,7 @@ run: build
 rebuild: clean build
 
 test-backend-unittets: frb-generate
-    cd {{ backend_dir }} && cargo test -- --nocapture
+    cd {{ backend_api_dir }} && cargo test -- --nocapture
 
 test-frontend-widget-tests: build
     cd {{ frontend_dir }} && flutter test
