+++ patches/frontend/models.dart.patch
@@ -1,5 +1,5 @@
---- a/frontend/lib/backend/models.dart.unpatched
-+++ b/frontend/lib/backend/models.dart
+--- a/frontend/lib/backend_api/models.dart.unpatched
++++ b/frontend/lib/backend_api/models.dart
 @@ -5,5 +5,7 @@
 
  import 'frb_generated.dart';
